requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.16
django-appconf==1.0.6
django-ckeditor==6.7.2
django-environ==0.12.0
django-js-asset==3.0.1
django-safedelete==1.4.0
django-select2==8.2.1
django-simple-history==3.7.0
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
channels==4.0.0
azure-cognitiveservices-speech==1.25.0